Changes to generating of docker image, angular services, some updates to login GUI process

+++ src/main/java/sk/filo/plantdiary/config/ConfigProperties.java
@@ -21,7 +21,7 @@ public class ConfigProperties {
     private String jwtPublicKey;
 
     @NotNull
-    private Integer jwtExpirationTime;
+    private Long jwtExpirationTime;
 
     @NotNull
     private String serverEmail;
